Added comparison option (k = 1)

+++ Coords.h
@@ -1,5 +1,5 @@
 #pragma once
 
 struct Coords {
-    int x, y;
+    long double x, y;
 };
